[HIBIKE] Full implementation of async smart sensor protocol (#523)
* Cleaned up some code

* Draft for async write function

* Coalesce sensor data into single class

* Implement read thread functionality for protocol

* Parse incoming packets from serial

* Stop coroutines on transport close

* Add ability to terminate sensor protocol

* Allow sensor to register itself w/ hibike_process

* Partial impl of async hotplug

Now, SmartSensorProtocol automagically registers itself with
Hibike when it connects.

* Use aioprocessing library

* Remove unnecessary code and redundant imports

* Fix small bugs

* Fix bugs related to virtual devices

Essentially, we exclude existing serial ports from our
scan for new ones, but this didn't extend to virtual
devices, leading them to be added multiple times.

The other bug was that "connection_lost()" could
get called before a device was identified, triggering
a key error when we tried to take it out of the
devices map. This is now checked for.